disable dhclient from modifying the /etc/resolv.conf file on openstack/amazon instances
The dhclient in these environments is quite aggressive and overwrites the
nameservers we've deliberately chosen to use with google's nameservers. This
commit attempts to fix that.

The dhclient methodology for altering these things is particularly
unpleasant. We effectively redefine the functions that mess with this file to be
noops in the /etc/dhcp/dhclient-enter-hooks.d directory and then we are forced
to restart dhclient by shipping a script that tries to determine the correct PID
and arguments that it was running as before killing and restarting it with the
same arguments.

See debian bugs #681698, #712796 for further discussion about how to make this
less difficult

Change-Id: I51cf40cf98eaddcefd8180e157b6e3ca824173f0