Copy the current state of the platform documentation into the doc directory.
Originally I thought it would be better to add the leap_doc git repository as a
submodule, but I decided against that:

 . it requires that the user has to start off by initializing submodules,
 something that the leap_cli does for you

 . it would result in more up-to-date documentation than was targeted for this
 release

 . it would result in an unfortunate directory structure (doc/doc/platform).

For these reasons it seemed to me better to put a snapshot of our current
platform documentation into the doc directory right before release. This just
means a step in our release process of refreshing these docs once we have
reviewed them and updated the known-issues for this release.

Change-Id: Ib395ea30553772fd195dd50315f026a2576feedd