Adopt tapicero config file to tapicero > 0.2.0
In order to reuse the couch changes observer for key uploads I extracted it from tapicero and made the configuration more generic. This results in some changes to the config file format.

When deploying tapicero >= 0.2.0 this config file format should be used instead.

Also set log_level to debug by default. We can reduce this later but it might be useful for now.