Switch away from site_config::sshd and instead just include site_sshd
The existing site_config::sshd had a non-functioning 'include sshd' line
in it that was not doing what was expected (this was supposed to include
the sshd module, but due to scoping was including itself).

It seemed better to eliminate some of the unused pieces and consolidate
into one config location.

Change-Id: I79dd904e696ca646180a09abbb03b5361dfc8ab9