Ignore bigcouch conflict errors, mainly coming from tapicero creating new users (Feature #6481)
There are potentially many tapicero daemons running, and they all try
to do the same thing at the same time. It is basically designed to
create race conditions. All tapicero daemons try to create the user db
at the same time. Only one of them wins the race and actually creates
it.

We need to fix this later (see https://leap.se/code/issues/6480) but
for now, we ignore them because conflict errors should be handled by the
applictation anyway.

Change-Id: I91095b1901d238e3d199954ba3716023d3fd49c1