Use augeas instead of file_line to configure entries in /etc/check_mk/mrpe.cfg (Bug #6788)
We used file_line before, but when the some check parameters change, a
new line would be added, leaving the old line there, resulting in two
checks with the same name but with different parameters.

Augeas can handle this better, but it is important to use 'rm' to
remove all old lines with different parameters before adding the
new line.

Change-Id: Iad69dfd20f487a16d372a4f4a4bc53299f9e4a66