[bug] Fix missing dependency (tapicero leftovers)
We need to remove local check-mk-agent checks on the tapicero
nodes, and want to notify the monitoring server to re-inventarize the
local checks.
This doesn't work when both services run on different hosts, it will
fail with:

    Could not find dependent Exec[check_mk-refresh] for Tidy[checkmk_logwatch_spool]

So i remove the notifies, because we will re-inventarize of local checks
by a daily cronjob anyway, see #6873.

...

- Resolves: #XYZ
- Related: #XYZ
- Documentation: #XYZ
- Releases: XYZ