[bug] do not fail if no requests/srp dependencies present
client_side_db helper has some extra functions that can be useful in the
future for further tests, but right now it shouldn't depend on requests,
since it already get all the relevant soledad info passed as arguments.

for the same reason, it doesn't need to depend on srp, since we pass the
token.