refresh_stunnel sometimes doesn't run (#8168).
It turns out that in some corner-cases, the script is not called:

 (1) start the deploy, create files in /var/lib/puppet/stunnel4/config
 (2) halt puppet before apply finishes
 (3) re-run deploy

in this scenario, next time you run deploy, refresh_stunnel will never
get called to populate /etc/stunnel, because the files in
/var/lib/puppet/stunnel4/config haven't changed.

This problem can be really confusing when it happens.

To fix this, we just run refresh_stunnel every, it is pretty fast and
the script has more complete logic for what to do than puppet, which has
only an asymmetrical view on the situation.

Change-Id: I9e5fad1d081c2fe07f3ac8f07cfb87d86b88f7c9