Stop tor from restarting on every deploy (#8211).
We were creating the hidden service name without a newline, and then tor
would be restarted and change the hidden service hostname file to have a
newline, which would then require that the next deploy would change that
file to not have a newline again.

This fixes that problem by making the hostname have a newline so it
matches what tor wants.

Change-Id: I38f450684d557cf943ec94f2f8e19cda3aefdf66