Set TCP_NODELAY option for couchdb (#8264)
Mochiweb in couchdb by default sets the TCP socket option SO_NODELAY to
false. This means that small data sent to the TCP socket, like the reply
to a document write request (or reading a very small document), will not
be sent immediately to the network - TCP will buffer it for a while
hoping that it will be asked to send more data through the same socket
and then send all the data at once for increased performance.

Setting this increases the couchdb speed significantly.

Change-Id: Ib493ef061ff62c9bdee501e44ce2b55990fe14b7