Do a dist-upgrade in 'node init'.
Starting with a recent debian stable point release update, it is
possible that the system is in an inconsistent library state. For
example, puppet could not be run because the libraries on the system
were not the ones that the puppet package was built against. So that
means that deploys could not happen until we've dont a dist-upgrade.