CI: Fix caching between jobs
fog-aws gem was not installed so Job #19895 failed for b566104d
The reason was that caching was configured wrongly so files got
cached vertically for each job.
After re-configuring the gitlab-runner cache on beluga, caches
were lost and resulted in failing builds.
This commit configures caching in a way that we define one global
cache for all platform pipelines.

Resolves: #8872