webapp: alert on 409 responses
They might be meaningful response codes for some scenarios. But
so far we are not conciously sending them out. If they occur that
is because we handed them down from couch. So we might want to
fix the underlying issue. Couch 409s should be caught by the webapp
and handled there.