webapp: update identity design doc for katzenpost
As part of webapp#8815 we want to retrieve multiple types of
keys through nickserver. This requires the new couchdb view
on Identities this commit provides.

The webapp and platform version of couchdb design docs need to
be in sync. Therefore this mr should be merged at the same time
as webapp!61 .